feat: refactor for reusing localstorage and remotestorage column state hooks
