+++ src/helpers/o2m-columnStorageHelper.ts
@@ -1,0 +1,13 @@
+export type One2manyTreeDataForHash = {
+  parentViewId?: number;
+  treeViewId?: number;
+  one2ManyFieldName: string;
+};
+
+export type O2mDataForHashWithModel = One2manyTreeDataForHash & {
+  model: string;
+};
+
+export const getKey = (dataForHash: O2mDataForHashWithModel) => {
+  return `columnState-${dataForHash.parentViewId}-${dataForHash.treeViewId}-${dataForHash.one2ManyFieldName}-${dataForHash.model}`;
+};
