+++ src/widgets/base/one2many/useTreeColumnRemoteStorage.ts
@@ -1,16 +1,10 @@
 import { ColumnState } from "@gisce/react-formiga-table";
-import { One2manyTreeDataForHash } from "./One2manyTree";
 import { useDeepCompareCallback } from "use-deep-compare";
 import ConnectionProvider from "@/ConnectionProvider";
 import { useNetworkRequest } from "@/hooks/useNetworkRequest";
-import { getKey } from "./useOne2manyColumnStorage";
 import { useEffect } from "react";
 
-export type DataForHashWithModel = One2manyTreeDataForHash & { model: string };
-
-export const useOne2manyColumnRemoteStorage = (
-  dataForHash: DataForHashWithModel,
-) => {
+export const useTreeColumnRemoteStorage = (key: string) => {
   useEffect(() => {
     return () => {
       cancelReadRequest();
@@ -31,13 +25,13 @@ export const useOne2manyColumnRemoteStorage = (
     ColumnState[] | undefined
   > => {
     const state = await read({
-      key: getKey(dataForHash),
+      key,
     });
     if (state === false) {
       throw new Error("Empty column state");
     }
     return state;
-  }, [dataForHash]);
+  }, [key]);
 
   const updateColumnState = useDeepCompareCallback(
     async (state: ColumnState[]) => {
@@ -50,11 +44,11 @@ export const useOne2manyColumnRemoteStorage = (
       );
 
       return save({
-        key: getKey(dataForHash),
+        key,
         preferences: stateWithoutNulls,
       });
     },
-    [dataForHash],
+    [key],
   );
 
   return { getColumnState, updateColumnState };
