+++ src/widgets/base/one2many/useTreeColumnStorageFetch.ts
@@ -1,18 +1,13 @@
 import { ColumnState } from "@gisce/react-formiga-table";
-import { One2manyTreeDataForHash } from "./One2manyTree";
-import { useOne2manyColumnStorage } from "./useOne2manyColumnStorage";
 import { useCallback, useEffect, useRef, useState } from "react";
+import { useTreeColumnStorage } from "./useTreeColumnStorage";
 
-export type DataForHashWithModel = One2manyTreeDataForHash & { model: string };
-
-export const useOne2manyColumnStorageFetch = (
-  dataForHash: DataForHashWithModel,
-) => {
+export const useTreeColumnStorageFetch = (key: string) => {
   const [loading, setLoading] = useState(true);
   const columnState = useRef<ColumnState[] | undefined>(undefined);
 
   const { getColumnState: getColumnStateInternal, updateColumnState } =
-    useOne2manyColumnStorage(dataForHash);
+    useTreeColumnStorage(key);
 
   useEffect(() => {
     const fetchColumnState = async () => {
